Fixed build error

+++ lib/utils/utils.dart
@@ -77,8 +77,7 @@ class Utils {
       return value;
     }
 
-    return Decimal.parse((value / BigInt.from(10).pow(unit)).toString())
-        .toString();
+    return Decimal.parse((value / BigInt.from(10).pow(unit)).toString()).toString();
   }
 
   formatUnitsByEtherUnit(
@@ -96,17 +95,17 @@ class Utils {
     String value, {
     int unit = 18,
   }) {
-    return (Decimal.parse(value) * Decimal.ten.pow(unit)).toBigInt();
+    return (Decimal.parse(value) * Decimal.ten.pow(unit).toDecimal()).toBigInt();
   }
 
   parseUnitsByEtherUnit(
     String value, {
     EtherUnit unit = EtherUnit.ether,
   }) {
-    return (Decimal.parse(value) * Decimal.ten.pow(_pows[unit]!)).toBigInt();
+    return (Decimal.parse(value) * Decimal.ten.pow(_pows[unit]!).toDecimal()).toBigInt();
   }
 
   BigInt parseEther(String value) {
-    return (Decimal.parse(value) * Decimal.ten.pow(18)).toBigInt();
+    return (Decimal.parse(value) * Decimal.ten.pow(18).toDecimal()).toBigInt();
   }
 }
